refactor: move files from :data:coreandroid to :data:core
