+++ app/build.gradle.kts
@@ -66,7 +66,6 @@ dependencies {
 	implementation(projects.data.core)
 	implementation(projects.data.settings)
 	implementation(projects.data.game)
-	implementation(projects.data.coreandroid)
 
 	implementation(libs.bundles.koin.compose)
 	implementation(libs.androidx.core.ktx)
