+++ data/core/build.gradle.kts
@@ -3,6 +3,7 @@ plugins {
 }
 
 dependencies {
-	implementation(libs.androidx.datastorePreferences)
 	implementation(libs.koin.android)
+	implementation(libs.androidx.datastorePreferences)
+	implementation(libs.androidx.proto.datastore)
 }
