+++ data/core/src/main/java/com/example/data/core/Module.kt
@@ -5,6 +5,8 @@ import androidx.datastore.preferences.core.Preferences
 import com.example.data.core.preferences.DataStorePreferenceStorage
 import com.example.data.core.preferences.PreferenceStorage
 import com.example.data.core.preferences.createDataStore
+import com.example.data.core.proto.ProtoStorageFactory
+import com.example.data.core.proto.ProtoStorageFactoryImpl
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
@@ -15,4 +17,6 @@ val dataCoreModule = module {
 	singleOf(::DataStorePreferenceStorage) {
 		bind<PreferenceStorage>()
 	}
+
+	single<ProtoStorageFactory> { ProtoStorageFactoryImpl(androidContext()) }
 }
