--- data/coreandroid/build.gradle.kts
@@ -1,10 +1,0 @@
-plugins {
-	id("DataModuleConvention")
-}
-
-dependencies {
-	implementation(projects.data.core)
-
-	implementation(libs.androidx.proto.datastore)
-	implementation(libs.koin.android)
-}
