--- data/coreandroid/src/main/java/com/example/data/coreandroid/CoreAndroidModule.kt
@@ -1,10 +1,0 @@
-package com.example.data.coreandroid
-
-import com.example.data.core.proto.ProtoStorageFactory
-import org.koin.android.ext.koin.androidContext
-import org.koin.dsl.module
-
-val coreAndroidModule =
-	module {
-		single<ProtoStorageFactory> { ProtoStorageFactoryImpl(androidContext()) }
-	}
