+++ settings.gradle.kts
@@ -31,7 +31,6 @@ include(":sudoku-core")
 include(":data:core")
 include(":data:settings")
 include(":data:game")
-include(":data:coreandroid")
 include(":domain:game")
 include(":domain:creator")
 include(":domain:core")
